added new CustomLink component and move on CustomLink function
