+++ src/components/index.jsx
@@ -2,3 +2,4 @@ export { default as Navbar } from "./Navbar";
 export { default as Footer } from "./Footer";
 export { default as DropDownMenu } from "./DropDownMenu";
 export { default as Router } from "./Routes";
+export { default as CustomLink } from "./CustomLink";
